added torch version of tabular neural network to complement mxnet (#1489)
* refactor torch/mxnet split after rebase

* prerebase quantile

* prerebase quantile all files added

* post rebase with master and bugfix

* updated model-name in example

* fix unit test

* delete extra file

* import bug

* Fix predict_proba metric crash for Torch NN during fit

* Fix seed to 0 by default for Torch and MXNet NN

* fix MXNet NN imports

* fix pred_proba probabilities

* print architecture and epoch numbers

* optimize boolean support

* code cleanup

* code cleanup

* code cleanup

* change updates to epochs

* further updates

* code cleanup

* add todo

* only NN_TORCH in default config for testing

* switch usage of MXNET to TORCH
