+++ tabular/src/autogluon/tabular/models/tabular_nn/mxnet/embednet.py
@@ -2,6 +2,8 @@ import numpy as np
 import mxnet as mx
 from mxnet import nd, gluon
 
+from ..utils.nn_architecture_utils import get_embed_sizes
+
 
 class NumericBlock(gluon.HybridBlock):
     """ Single Dense layer that jointly embeds all numeric and one-hot features """
@@ -80,15 +82,13 @@ class EmbedNet(gluon.Block): # TODO: hybridize?
             self.from_logits = False
             self.has_vector_features = train_dataset.has_vector_features()
             self.has_embed_features = train_dataset.num_embed_features() > 0
-            self.has_language_features = train_dataset.num_language_features() > 0
             if self.has_embed_features:
                 num_categs_per_feature = train_dataset.getNumCategoriesEmbeddings()
-                embed_dims = getEmbedSizes(train_dataset, params, num_categs_per_feature)
+                embed_dims = get_embed_sizes(train_dataset, params, num_categs_per_feature)
         else: # Ignore train_dataset, params, etc. Recreate architecture based on description:
             self.architecture_desc = architecture_desc
             self.has_vector_features = architecture_desc['has_vector_features']
             self.has_embed_features = architecture_desc['has_embed_features']
-            self.has_language_features = architecture_desc['has_language_features']
             self.from_logits = architecture_desc['from_logits']
             num_net_outputs = architecture_desc['num_net_outputs']
             params = architecture_desc['params']
@@ -103,9 +103,6 @@ class EmbedNet(gluon.Block): # TODO: hybridize?
             self.embed_blocks = gluon.nn.HybridSequential()
             for i in range(len(num_categs_per_feature)):
                 self.embed_blocks.add(EmbedBlock(embed_dims[i], num_categs_per_feature[i]))
-        if self.has_language_features:
-            self.text_block = None
-            raise NotImplementedError("text data cannot be handled")
         if params['network_type'] == 'feedforward':
             self.output_block = FeedforwardBlock(params, num_net_outputs)
         elif params['network_type'] == 'widedeep':
@@ -134,15 +131,12 @@ class EmbedNet(gluon.Block): # TODO: hybridize?
         if architecture_desc is None: # Save Architecture description
             self.architecture_desc = {'has_vector_features': self.has_vector_features, 
                                   'has_embed_features': self.has_embed_features,
-                                  'has_language_features': self.has_language_features,
                                   'params': params, 'num_net_outputs': num_net_outputs,
                                   'from_logits': self.from_logits}
             if self.has_embed_features:
                 self.architecture_desc['num_categs_per_feature'] = num_categs_per_feature
                 self.architecture_desc['embed_dims'] = embed_dims
-            if self.has_language_features:
-                self.architecture_desc['text_TODO'] = None # TODO: store text architecture
-    
+
     def forward(self, data_batch):
         if self.has_vector_features:
             numerical_data = data_batch['vector'] # NDArray
@@ -199,13 +193,6 @@ class EmbedNet(gluon.Block): # TODO: hybridize?
                 input_activations = embed_activations
             else:
                 input_activations = nd.concat(embed_activations, input_activations)
-        if self.has_language_features:
-            language_data = data_batch['language']
-            language_activations = self.text_block(language_data) # TODO: create block to embed text fields
-            if (not self.has_vector_features) and (not self.has_embed_features):
-                input_activations = language_activations
-            else:
-                input_activations = nd.concat(language_activations, input_activations)
         if self.y_constraint is None:
             return self.output_block(input_activations)
         else:
@@ -224,43 +211,3 @@ class EmbedNet(gluon.Block): # TODO: hybridize?
                 print("nd.sigmoid(unscaled_pred).shape", nd.sigmoid(unscaled_pred).shape)
                 """
                 return nd.sigmoid(unscaled_pred) * self.y_span + self.y_lower
-
-
-""" OLD 
-    def _create_embednet_from_architecture(architecture_desc):
-        # Recreate network architecture based on provided description
-        self.architecture_desc = architecture_desc
-        self.has_vector_features = architecture_desc['has_vector_features']
-        self.has_embed_features = architecture_desc['has_embed_features']
-        self.has_language_features = architecture_desc['has_language_features']
-        self.from_logits = architecture_desc['from_logits']
-        num_net_outputs = architecture_desc['num_net_outputs']
-        params = architecture_desc['params']
-        if self.has_vector_features:
-            self.numeric_block = NumericBlock(params)
-        if self.has_embed_features:
-            self.embed_blocks = gluon.nn.HybridSequential()
-            num_categs_per_feature = architecture_desc['num_categs_per_feature']
-            embed_dims = architecture_desc['embed_dims']
-            for i in range(len(num_categs_per_feature)):
-                self.embed_blocks.add(EmbedBlock(embed_dims[i], num_categs_per_feature[i]))
-        if self.has_language_features:
-            self.text_block = architecture_desc['text_TODO']
-        if 
-        self.output_block = FeedforwardBlock(params, num_net_outputs) # TODO
-        self.from_logits = False
-"""
-
-
-def getEmbedSizes(train_dataset, params, num_categs_per_feature):  
-    """ Returns list of embedding sizes for each categorical variable.
-        Selects this adaptively based on training_datset.
-        Note: Assumes there is at least one embed feature.
-    """
-    max_embedding_dim = params['max_embedding_dim']
-    embed_exponent = params['embed_exponent']
-    size_factor = params['embedding_size_factor']
-    embed_dims = [int(size_factor*max(2, min(max_embedding_dim, 
-                                      1.6 * num_categs_per_feature[i]**embed_exponent)))
-                   for i in range(len(num_categs_per_feature))]
-    return embed_dims
