+++ tabular/src/autogluon/tabular/models/xgboost/xgboost_utils.py
@@ -5,7 +5,7 @@ from sklearn.base import BaseEstimator, TransformerMixin
 from sklearn.preprocessing import OneHotEncoder
 from autogluon.core.constants import BINARY, MULTICLASS, REGRESSION, SOFTCLASS
 
-from ..tabular_nn.categorical_encoders import OneHotMergeRaresHandleUnknownEncoder
+from ..tabular_nn.utils.categorical_encoders import OneHotMergeRaresHandleUnknownEncoder
 
 
 _ag_to_xgbm_metric_dict = {
