+++ tabular/src/autogluon/tabular/trainer/model_presets/presets.py
@@ -9,7 +9,7 @@ from autogluon.core.trainer.utils import process_hyperparameters
 
 from .presets_custom import get_preset_custom
 from ...models import LGBModel, CatBoostModel, XGBoostModel, RFModel, XTModel, KNNModel, LinearModel,\
-    TabularNeuralNetModel, TabularNeuralQuantileModel, NNFastAiTabularModel, FastTextModel, TextPredictorModel, \
+    TabularNeuralNetMxnetModel, TabularNeuralNetTorchModel, NNFastAiTabularModel, FastTextModel, TextPredictorModel, \
     ImagePredictorModel, VowpalWabbitModel
 from ...models.tab_transformer.tab_transformer_model import TabTransformerModel
 
@@ -25,7 +25,8 @@ DEFAULT_MODEL_PRIORITY = dict(
     FASTAI=50,
     XGB=40,
     LR=30,
-    NN=20,
+    NN_TORCH=25,
+    NN_MXNET=20,
     VW=10,
     FASTTEXT=0,
     AG_TEXT_NN=0,
@@ -43,7 +44,7 @@ PROBLEM_TYPE_MODEL_PRIORITY = {
 
 DEFAULT_SOFTCLASS_PRIORITY = dict(
     GBM=100,
-    NN=90,
+    NN_MXNET=90,
     RF=80,
     CAT=60,
     custom=0,
@@ -51,7 +52,7 @@ DEFAULT_SOFTCLASS_PRIORITY = dict(
 
 DEFAULT_CUSTOM_MODEL_PRIORITY = 0
 
-DEFAULT_QUANTILE_MODEL = ['RF', 'XT', 'FASTAI', 'QNN', 'ENS_WEIGHTED'] # TODO: OTHERS will be added
+DEFAULT_QUANTILE_MODEL = ['RF', 'XT', 'FASTAI', 'NN_TORCH', 'ENS_WEIGHTED']  # TODO: OTHERS will be added
 
 MODEL_TYPES = dict(
     RF=RFModel,
@@ -60,8 +61,8 @@ MODEL_TYPES = dict(
     GBM=LGBModel,
     CAT=CatBoostModel,
     XGB=XGBoostModel,
-    NN=TabularNeuralNetModel,
-    QNN=TabularNeuralQuantileModel,
+    NN_MXNET=TabularNeuralNetMxnetModel,
+    NN_TORCH=TabularNeuralNetTorchModel,
     LR=LinearModel,
     FASTAI=NNFastAiTabularModel,
     TRANSF=TabTransformerModel,
@@ -80,8 +81,8 @@ DEFAULT_MODEL_NAMES = {
     LGBModel: 'LightGBM',
     CatBoostModel: 'CatBoost',
     XGBoostModel: 'XGBoost',
-    TabularNeuralNetModel: 'NeuralNetMXNet',
-    TabularNeuralQuantileModel: 'QuantileNeuralNet',
+    TabularNeuralNetMxnetModel: 'NeuralNetMXNet',
+    TabularNeuralNetTorchModel: 'NeuralNetTorch',
     LinearModel: 'LinearModel',
     NNFastAiTabularModel: 'NeuralNetFastAI',
     TabTransformerModel: 'Transformer',
@@ -155,13 +156,17 @@ def get_preset_models(path, problem_type, eval_metric, hyperparameters,
     hp_level = hyperparameters[level_key]
     model_cfg_priority_dict = defaultdict(list)
     model_type_list = list(hp_level.keys())
+    if 'NN' in model_type_list:
+        raise ValueError("'NN' model has been deprecated. Please specify 'NN_MXNET' or 'NN_TORCH' in its place (Tabular Neural Networks implemented in different backend frameworks).")
     for model_type in model_type_list:
-        if problem_type == QUANTILE and model_type not in DEFAULT_QUANTILE_MODEL:
-            if model_type == 'NN' and 'QNN' in DEFAULT_QUANTILE_MODEL:
-                model_type = 'QNN'
-                hp_level['QNN'] = hp_level.pop('NN')
-            else:
-                continue
+        if problem_type == QUANTILE:
+            if model_type not in DEFAULT_QUANTILE_MODEL:
+                if model_type == 'NN_MXNET' and 'NN_TORCH' in DEFAULT_QUANTILE_MODEL and 'NN_TORCH' not in model_type_list:
+                    model_type = 'NN_TORCH'
+                    hp_level['NN_TORCH'] = hp_level.pop('NN_MXNET')
+                    logger.log(15, "Quantile regression must use NN_TORCH instead of NN_MXNET, switching NN_MXET -> NN_TORCH.")
+                else:
+                    continue
         models_of_type = hp_level[model_type]
         if not isinstance(models_of_type, list):
             models_of_type = [models_of_type]
@@ -338,7 +343,7 @@ def model_factory(
 def get_preset_models_softclass(hyperparameters, invalid_model_names: list = None, **kwargs):
     # TODO v0.1: This import depends on mxnet, consider refactoring to avoid mxnet
     from autogluon.core.metrics.softclass_metrics import soft_log_loss
-    model_types_standard = ['GBM', 'NN', 'CAT', 'ENS_WEIGHTED']
+    model_types_standard = ['GBM', 'NN_MXNET', 'NN_TORCH', 'CAT', 'ENS_WEIGHTED']
     hyperparameters = copy.deepcopy(hyperparameters)
 
     hyperparameters_standard = {key: hyperparameters[key] for key in hyperparameters if key in model_types_standard}
@@ -356,7 +361,7 @@ def get_preset_models_softclass(hyperparameters, invalid_model_names: list = Non
                                                hyperparameters=hyperparameters_standard,
                                                default_priorities=DEFAULT_SOFTCLASS_PRIORITY, invalid_model_names=invalid_model_names, **kwargs)
     if len(models) == 0:
-        raise ValueError("At least one of the following model-types must be present in hyperparameters: ['GBM','CAT','NN','RF'], "
+        raise ValueError("At least one of the following model-types must be present in hyperparameters: ['GBM','CAT','NN_MXNET','RF'], "
                          "These are the only supported models for softclass prediction problems. "
                          "Softclass problems are also not yet supported for fit() with per-stack level hyperparameters.")
     for model in models:
