+++ tabular/tests/unittests/models/test_tabular_nn.py
@@ -1,10 +1,10 @@
 
-from autogluon.tabular.models.tabular_nn.tabular_nn_model import TabularNeuralNetModel
+from autogluon.tabular.models.tabular_nn.torch.tabular_nn_torch import TabularNeuralNetTorchModel
 
 
 def test_tabular_nn_binary(fit_helper):
     fit_args = dict(
-        hyperparameters={TabularNeuralNetModel: {}},
+        hyperparameters={TabularNeuralNetTorchModel: {}},
     )
     dataset_name = 'adult'
     fit_helper.fit_and_validate_dataset(dataset_name=dataset_name, fit_args=fit_args)
@@ -12,7 +12,7 @@ def test_tabular_nn_binary(fit_helper):
 
 def test_tabular_nn_multiclass(fit_helper):
     fit_args = dict(
-        hyperparameters={TabularNeuralNetModel: {}},
+        hyperparameters={TabularNeuralNetTorchModel: {}},
     )
     dataset_name = 'covertype'
     fit_helper.fit_and_validate_dataset(dataset_name=dataset_name, fit_args=fit_args)
@@ -20,7 +20,7 @@ def test_tabular_nn_multiclass(fit_helper):
 
 def test_tabular_nn_regression(fit_helper):
     fit_args = dict(
-        hyperparameters={TabularNeuralNetModel: {}},
+        hyperparameters={TabularNeuralNetTorchModel: {}},
         time_limit=10,  # TabularNN trains for a long time on ames
     )
     dataset_name = 'ames'
